@@ -11,21 +11,21 @@
     "@types/jest": "^27.4.0",
     "@types/node": "^16.11.22",
     "@types/object-hash": "^2.2.1",
-    "@types/react": "^16.8.5",
+    "@types/react": "^19.2.7",
     "@types/react-beautiful-dnd": "^13.1.2",
-    "@types/react-dom": "^16.8.5",
+    "@types/react-dom": "^19.2.3",
     "@types/react-pdf": "^5.0.0",
     "@uiw/react-codemirror": "~4.7.0",
-    "bootstrap": "^4.6.1",
+    "bootstrap": "^5.3.8",
     "classnames": "^2.3.1",
     "fast-json-patch": "^3.1.1",
     "he": "^1.2.0",
     "js-cookie": "^3.0.1",
     "katex": "^0.15.1",
     "object-hash": "^3.0.0",
-    "react": "^16.8.5",
+    "react": "^19.2.3",
     "react-beautiful-dnd": "^13.1.0",
-    "react-dom": "16.8.5",
+    "react-dom": "^19.2.3",
     "react-drag-drop-files": "^2.3.7",
     "react-error-boundary": "^3.1.4",
     "react-pdf": "^5.0.0",
@@ -34,11 +34,11 @@
     "react-scripts": "5.0.0",
     "react-select": "^5.3.2",
     "react-split": "^2.0.14",
-    "reactstrap": "^8.10.1",
+    "reactstrap": "^9.2.3",
     "remarkable": "^2.0.1",
     "swr": "^2.0.3",
     "ts-jest": "^29.2.5",
-    "typescript": "^5.8.3",
+    "typescript": "^5.9.3",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
